Remap backspace button in matplotlib to do nothing.
By default backspace in mpl restore previous zoom view. This is in
conflict with 'plt.ginput(-1)' that makes use of backspace to trash
the last acquired point.
It is of key importance the possibility to use backspace for ginput
because when the zoom (rectangle) is used, the corner of the rectangle
is acquired and must be removed with backspace.

Changes:
    modified:   PyDigitizer.py